requirements.txt
numpy~=2.2.4
Flask~=3.1.0
Flask-SocketIO~=5.5.1
pillow~=11.2.0